+++ src/main/java/EdTech/Course/dto/CourseDto.java
@@ -3,13 +3,14 @@ package EdTech.Course.dto;
 import EdTech.Course.model.CourseMaterial;
 import EdTech.Course.model.Enrollment;
 import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.NoArgsConstructor;
+import lombok.AllArgsConstructor;
+
 import java.util.List;
 
-@Getter
-@Setter
 @Data
+@NoArgsConstructor
+@AllArgsConstructor
 public class CourseDto {
     private String name;
     private String description;