+++ src/main/java/EdTech/Course/dto/Payment.java
@@ -1,8 +1,12 @@
 package EdTech.Course.dto;
 
 import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.AllArgsConstructor;
 
 @Data
+@NoArgsConstructor
+@AllArgsConstructor
 public class Payment {
     private Long userId;
     private Long courseId;