+++ src/main/java/EdTech/Course/dto/ResponseMessage.java
@@ -1,12 +1,12 @@
 package EdTech.Course.dto;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-@Getter
-@Setter
 public class ResponseMessage {
     private String message;
 }